+++ packages/components/components/elvis-modal/src/react/declarations.d.ts
@@ -1,4 +1,2 @@
 declare module '*.css';
 declare module '*.scss';
-declare module '@elvia/elvis-colors';
-declare module '@elvia/elvis-typography';