+++ packages/components/components/elvis-card/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Elvia Card Changelog
 
+## 1.0.1 (07.01.22)
+
+### New feature
+
+- Updated elvis-typography use method
+
 ## 1.0.0 (03.01.22)
 
 ### New feature