+++ packages/elvis-colors/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Elvia Colors Changelog
 
+## 1.2.0 (07.11.22)
+
+### New feature
+
+- Added JSDoc documentation and better Typescript declarations.
+
 ## 1.1.0 (29.11.21)
 
 ### New feature