+++ packages/elvis-typography/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Elvia Typography Changelog
 
+## 1.2.0 (07.11.22)
+
+### New feature
+
+- Added JSDoc documentation and better Typescript declarations.
+
 ## 1.1.1 (17.11.21)
 
 ### Bug fix